requirements.txt
clickhouse-connect
scrapy
scrapeops-scrapy
google-generativeai
beautifulsoup4
minio
python-dotenv
fitz
requests
PyMuPDF
pandas
python-dotenv
boto3